+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Delete itinerary bug fixed
+- Search course bug fixed
 
 ## [1.3.0] - 27/05/2020